Move tools to a separate project, reduce dependency conflicts with code
