+++ internal/tools/go.mod
@@ -1,0 +1,15 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/internal/tools
+
+go 1.15
+
+require (
+	github.com/client9/misspell v0.3.4
+	github.com/golangci/golangci-lint v1.35.2
+	github.com/google/addlicense v0.0.0-20200906110928-a0294312aa76
+	github.com/jstemmer/go-junit-report v0.9.1
+	github.com/pavius/impi v0.0.3
+	github.com/tcnksm/ghr v0.13.0
+	go.opentelemetry.io/collector/cmd/issuegenerator v0.0.0-20210107171527-a263ced270b9
+	go.opentelemetry.io/collector/cmd/mdatagen v0.0.0-20210107171527-a263ced270b9
+	honnef.co/go/tools v0.0.1-2020.1.6
+)
